Refactor submit command action into handlers/submit
This PR extracts the `submit` command action handler into a stand alone handler
in `handlers/submit.go` to match the other actions.

  * The related specs from `main.go` have been moved to
`submit_test.go`.
  * `Version` and `UserAgent` was moved out of `main` and into
`config` for this update to work. There may be a better place for these
constants / variables.
  * `msgPleaseAuthenticate` is currently duplicated in both `main` and in
  `submit`. Do you have a preferred way to share these messages?
