+++ handlers/submit_test.go
@@ -1,0 +1,126 @@
+package handlers
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"testing"
+	"net/http"
+	"net/http/httptest"
+	"strings"
+
+	"github.com/exercism/cli/config"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestSubmitWithKey(t *testing.T) {
+	server := httptest.NewServer(http.HandlerFunc(submitHandler))
+	defer server.Close()
+
+	var code = []byte("My source code\n")
+	c := &config.Config{
+		Hostname: server.URL,
+		APIKey:   "myAPIKey",
+	}
+	response, err := SubmitAssignment(c, "ruby/bob/bob.rb", code)
+	assert.NoError(t, err)
+
+	assert.Equal(t, response.Status, "saved")
+	assert.Equal(t, response.Language, "ruby")
+	assert.Equal(t, response.Exercise, "bob")
+	assert.Equal(t, response.SubmissionPath, "/username/ruby/bob")
+}
+
+func TestSubmitWithIncorrectKey(t *testing.T) {
+	server := httptest.NewServer(http.HandlerFunc(submitHandler))
+	defer server.Close()
+
+	c := &config.Config{
+		Hostname: server.URL,
+		APIKey:   "myWrongAPIKey",
+	}
+
+	var code = []byte("My source code\n")
+	_, err := SubmitAssignment(c, "ruby/bob/bob.rb", code)
+
+	assert.Error(t, err)
+}
+
+var submitHandler = func(rw http.ResponseWriter, r *http.Request) {
+	pathMatches := r.URL.Path == "/api/v1/user/assignments"
+	methodMatches := r.Method == "POST"
+	if !(pathMatches && methodMatches) {
+		rw.WriteHeader(http.StatusNotFound)
+		return
+	}
+
+	userAgentMatches := strings.HasPrefix(r.Header.Get("User-Agent"), fmt.Sprintf("github.com/exercism/cli v%s", config.Version))
+
+	if !userAgentMatches {
+		fmt.Printf("User agent mismatch: %s\n", r.Header.Get("User-Agent"))
+		rw.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	body, err := ioutil.ReadAll(r.Body)
+	r.Body.Close()
+
+	if err != nil {
+		rw.WriteHeader(http.StatusInternalServerError)
+		fmt.Printf("Reading body error: %s\n", err)
+		return
+	}
+
+	type Submission struct {
+		Key  string
+		Code string
+		Path string
+	}
+
+	submission := Submission{}
+
+	err = json.Unmarshal(body, &submission)
+	if err != nil {
+		fmt.Printf("Unmarshalling error: %v, Body: %s\n", err, body)
+		rw.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	if submission.Key != "myAPIKey" {
+		rw.WriteHeader(http.StatusForbidden)
+		rw.Header().Set("Content-Type", "application/json")
+		fmt.Fprintf(rw, `{"error": "Unable to identify user"}`)
+		return
+	}
+
+	code := submission.Code
+	filePath := submission.Path
+
+	codeMatches := string(code) == "My source code\n"
+	filePathMatches := filePath == "ruby/bob/bob.rb"
+
+	if !filePathMatches {
+		fmt.Printf("FilePathMismatch: File Path: %s\n", filePath)
+		rw.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	if !codeMatches {
+		fmt.Printf("Code Mismatch: Code: %v\n", code)
+		rw.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	rw.WriteHeader(http.StatusCreated)
+	rw.Header().Set("Content-Type", "application/json")
+
+	submitJSON := `
+{
+	"status":"saved",
+	"language":"ruby",
+	"exercise":"bob",
+	"submission_path":"/username/ruby/bob"
+}
+`
+	fmt.Fprintf(rw, submitJSON)
+}
