+++ main.go
@@ -1,15 +1,12 @@
 package main
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
-	"runtime"
-	"strings"
 
 	"github.com/codegangsta/cli"
 	"github.com/exercism/cli/config"
@@ -17,12 +14,6 @@ import (
 )
 
 const (
-	// Version is the current release of the command-line app.
-	// We try to follow Semantic Versioning (http://semver.org),
-	// but with the http://exercism.io app being a prototype, a
-	// lot of things get out of hand.
-	Version = "1.7.0"
-
 	msgPleaseAuthenticate = "You must be authenticated. Run `exercism configure --key=YOUR_API_KEY`."
 
 	descDebug     = "Outputs useful debug information."
@@ -38,12 +29,6 @@ const (
 	descLongRestore = "Restore will pull the latest revisions of exercises that have already been submitted. It will *not* overwrite existing files. If you have made changes to a file and have not submitted it, and you're trying to restore the last submitted version, first move that file out of the way, then call restore."
 )
 
-var (
-	// UserAgent is sent along as a header to HTTP requests that the
-	// CLI makes. This helps with debugging.
-	UserAgent = fmt.Sprintf("github.com/exercism/cli v%s (%s/%s)", Version, runtime.GOOS, runtime.GOARCH)
-)
-
 var FetchEndpoints = map[string]string{
 	"current":  "/api/v1/user/assignments/current",
 	"next":     "/api/v1/user/assignments/next",
@@ -51,22 +36,11 @@ var FetchEndpoints = map[string]string{
 	"exercise": "/api/v1/assignments",
 }
 
-var testExtensions = map[string]string{
-	"ruby":    "_test.rb",
-	"js":      ".spec.js",
-	"elixir":  "_test.exs",
-	"clojure": "_test.clj",
-	"python":  "_test.py",
-	"go":      "_test.go",
-	"haskell": "_test.hs",
-	"cpp":     "_test.cpp",
-}
-
 func main() {
 	app := cli.NewApp()
 	app.Name = "exercism"
 	app.Usage = "A command line tool to interact with http://exercism.io"
-	app.Version = Version
+	app.Version = config.Version
 	app.Flags = []cli.Flag{
 		cli.StringFlag{
 			Name:  "config, c",
@@ -133,59 +107,7 @@ func main() {
 			Name:      "submit",
 			ShortName: "s",
 			Usage:     descSubmit,
-			Action: func(ctx *cli.Context) {
-				if len(ctx.Args()) == 0 {
-					fmt.Println("Please enter a file name")
-					return
-				}
-
-				c, err := config.Read(ctx.GlobalString("config"))
-				if err != nil {
-					fmt.Println(err)
-					return
-				}
-
-				if !c.IsAuthenticated() {
-					fmt.Println(msgPleaseAuthenticate)
-					return
-				}
-
-				filename := ctx.Args()[0]
-
-				// Make filename relative to config.Dir.
-				absPath, err := absolutePath(filename)
-				if err != nil {
-					fmt.Printf("Couldn't find %v: %v\n", filename, err)
-					return
-				}
-				exDir := c.Dir + string(filepath.Separator)
-				if !strings.HasPrefix(absPath, exDir) {
-					fmt.Printf("%v is not under your exercism project path (%v)\n", absPath, exDir)
-					return
-				}
-				filename = absPath[len(exDir):]
-
-				if IsTest(filename) {
-					fmt.Println("It looks like this is a test, please submit a solution.")
-					return
-				}
-
-				code, err := ioutil.ReadFile(absPath)
-				if err != nil {
-					fmt.Printf("Error reading %v: %v\n", absPath, err)
-					return
-				}
-
-				response, err := SubmitAssignment(c, filename, code)
-				if err != nil {
-					fmt.Printf("There was an issue with your submission: %v\n", err)
-					return
-				}
-
-				fmt.Printf("For feedback on your submission visit %s%s%s\n",
-					c.API, "/submissions/", response.ID)
-
-			},
+			Action:    handlers.Submit,
 		},
 		{
 			Name:      "unsubmit",
@@ -217,29 +139,6 @@ func main() {
 		fmt.Errorf("%v", err)
 		os.Exit(1)
 	}
-}
-
-func absolutePath(path string) (string, error) {
-	path, err := filepath.Abs(path)
-	if err != nil {
-		return "", err
-	}
-	return filepath.EvalSymlinks(path)
-}
-
-type submitResponse struct {
-	ID             string `json:"id"`
-	Status         string `json:"status"`
-	Language       string `json:"language"`
-	Exercise       string `json:"exercise"`
-	SubmissionPath string `json:"submission_path"`
-	Error          string `json:"error"`
-}
-
-type submitRequest struct {
-	Key  string `json:"key"`
-	Code string `json:"code"`
-	Path string `json:"path"`
 }
 
 func FetchAssignments(c *config.Config, path string) ([]Assignment, error) {
@@ -300,7 +199,7 @@ func UnsubmitAssignment(c *config.Config) error {
 		return err
 	}
 
-	req.Header.Set("User-Agent", UserAgent)
+	req.Header.Set("User-Agent", config.UserAgent)
 
 	resp, err := http.DefaultClient.Do(req)
 	if err != nil {
@@ -330,53 +229,6 @@ func UnsubmitAssignment(c *config.Config) error {
 	}
 
 	return nil
-}
-func SubmitAssignment(c *config.Config, filePath string, code []byte) (*submitResponse, error) {
-	path := "api/v1/user/assignments"
-
-	url := fmt.Sprintf("%s/%s", c.API, path)
-
-	submission := submitRequest{Key: c.APIKey, Code: string(code), Path: filePath}
-	submissionJSON, err := json.Marshal(submission)
-	if err != nil {
-		return nil, err
-	}
-
-	req, err := http.NewRequest("POST", url, bytes.NewReader(submissionJSON))
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("User-Agent", UserAgent)
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		err = fmt.Errorf("Error posting assignment: [%v]", err)
-		return nil, err
-	}
-
-	body, err := ioutil.ReadAll(resp.Body)
-	resp.Body.Close()
-	if err != nil {
-		return nil, err
-	}
-
-	var r submitResponse
-	if resp.StatusCode != http.StatusCreated {
-		err = json.Unmarshal(body, &r)
-		if err != nil {
-			return nil, err
-		}
-		err = fmt.Errorf("Status: %d, Error: %v", resp.StatusCode, r)
-		return nil, err
-	}
-
-	err = json.Unmarshal(body, &r)
-	if err != nil {
-		return nil, fmt.Errorf("Error parsing API response: [%v]", err)
-	}
-	return &r, nil
 }
 
 type Assignment struct {
@@ -426,13 +278,4 @@ func FetchEndpoint(args []string) string {
 	}
 
 	return endpoint
-}
-
-func IsTest(filename string) bool {
-	for _, ext := range testExtensions {
-		if strings.LastIndex(filename, ext) > 0 {
-			return true
-		}
-	}
-	return false
 }
